codetime refactor
