+++ src/managers/PromptManager.ts
@@ -1,19 +1,15 @@
 import {commands, window} from 'vscode';
-
-export const SIGN_UP_LABEL = 'Sign up';
+import { SIGN_UP_LABEL } from '../Constants';
 
 export function showModalSignupPrompt(msg: string) {
   window
     .showInformationMessage(
       msg,
-      {
-        modal: true,
-      },
       SIGN_UP_LABEL
     )
     .then((selection: string | undefined) => {
       if (selection === SIGN_UP_LABEL) {
-        commands.executeCommand('');
+        commands.executeCommand('codetime.registerAccount');
       }
     });
 }
